Move StreetType to network (from parsing)
